requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
iniconfig==2.0.0
mysqlclient==2.2.6
packaging==24.2
pluggy==1.5.0
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==8.3.4
python-dotenv==1.0.1
sqlparse==0.5.3
